+++ lua/github-theme/init.lua
@@ -1,12 +1,18 @@
-local util = require("github-theme.util")
-local theme = require("github-theme.theme")
-local config_module = require("github-theme.config")
+local config = require('github-theme.config')
+local theme = require('github-theme.theme')
+local util = require('github-theme.util')
 
-local function setup(user_config)
-  if user_config then config_module.apply_configuration(user_config) end
+---@class gt.Init
+local init = {}
+
+---@param user_config gt.ConfigSchema
+init.setup = function(user_config)
+  if user_config then
+    config.apply_configuration(user_config)
+  end
 
   -- Load colorscheme
-  util.load(theme.setup(config_module.config))
+  util.load(theme.setup(config.schema))
 end
 
-return {setup = setup}
+return init