+++ lua/github-theme/util.lua
@@ -1,97 +1,101 @@
-local hsluv = require("github-theme.hsluv")
+local types = require('github-theme.types')
 
+---@class gt.Util
 local util = {}
 
-util.colors_name = ""
-util.colorsUsed = {}
-util.colorCache = {}
+util.colors_name = ''
 
-util.bg = "#000000"
-util.fg = "#ffffff"
+---@type table<number, gt.HexColor>
+util.used_color = {}
 
-local function hexToRgb(hex_str)
-  local hex = "[abcdef0-9][abcdef0-9]"
-  local pat = "^#(" .. hex .. ")(" .. hex .. ")(" .. hex .. ")$"
+---@type gt.HexColor
+util.bg = '#000000'
+
+---@type gt.HexColor
+util.fg = '#ffffff'
+
+---Convert HexColor to RGB
+---@param hex_str gt.HexColor
+---@return table<number, number>
+local hex_to_rgb = function(hex_str)
+  local hex = '[abcdef0-9][abcdef0-9]'
+  local pat = string.format('^#(%s)(%s)(%s)$', hex, hex, hex)
   hex_str = string.lower(hex_str)
 
-  assert(string.find(hex_str, pat) ~= nil,
-         "hex_to_rgb: invalid hex_str: " .. tostring(hex_str))
+  assert(string.find(hex_str, pat) ~= nil, 'hex_to_rgb: invalid hex_str: ' .. tostring(hex_str))
 
   local r, g, b = string.match(hex_str, pat)
-  return {tonumber(r, 16), tonumber(g, 16), tonumber(b, 16)}
+  return { tonumber(r, 16), tonumber(g, 16), tonumber(b, 16) }
 end
 
----@param fg string foreground color
----@param bg string background color
+---@param fg gt.HexColor foreground color
+---@param bg gt.HexColor background color
 ---@param alpha number number between 0 and 1. 0 results in bg, 1 results in fg
-function util.blend(fg, bg, alpha)
-  bg = hexToRgb(bg)
-  fg = hexToRgb(fg)
+---@return gt.HexColor
+util.blend = function(fg, bg, alpha)
+  bg = hex_to_rgb(bg)
+  fg = hex_to_rgb(fg)
 
-  local blendChannel = function(i)
+  local blend_channel = function(i)
     local ret = (alpha * fg[i] + ((1 - alpha) * bg[i]))
     return math.floor(math.min(math.max(0, ret), 255) + 0.5)
   end
 
-  return string.format("#%02X%02X%02X", blendChannel(1), blendChannel(2), blendChannel(3))
+  return string.format('#%02X%02X%02X', blend_channel(1), blend_channel(2), blend_channel(3))
 end
 
-function util.darken(hex, amount, bg)
+---@param hex gt.HexColor Color
+---@param amount number number between 0 and 1. 0 results in bg, 1 results in fg
+---@param bg gt.HexColor Background color
+---@return gt.HexColor
+util.darken = function(hex, amount, bg)
   return util.blend(hex, bg or util.bg, math.abs(amount))
 end
-function util.lighten(hex, amount, fg)
+
+---@param hex gt.HexColor Color
+---@param amount number number between 0 and 1. 0 results in bg, 1 results in fg
+---@param fg gt.HexColor Foreground color
+---@return gt.HexColor
+util.lighten = function(hex, amount, fg)
   return util.blend(hex, fg or util.fg, math.abs(amount))
 end
 
-function util.brighten(color, percentage)
-  local hsl = hsluv.hex_to_hsluv(color)
-  local larpSpace = 100 - hsl[3]
-  if percentage < 0 then larpSpace = hsl[3] end
-  hsl[3] = hsl[3] + larpSpace * percentage
-  return hsluv.hsluv_to_hex(hsl)
-end
-
-function util.invertColor(color)
-  if color ~= "NONE" then
-    local hsl = hsluv.hex_to_hsluv(color)
-    hsl[3] = 100 - hsl[3]
-    if hsl[3] < 40 then hsl[3] = hsl[3] + (100 - hsl[3]) * 0.3 end
-    return hsluv.hsluv_to_hex(hsl)
+---Dump unused colors
+---@param colors gt.ColorPalette
+util.debug = function(colors)
+  for name, color in pairs(colors) do
+    if type(color) == 'table' then
+      util.debug(color)
+    else
+      if util.used_color[color] == nil then
+        print(string.format('not used color: %s=%s', name, color))
+      end
+    end
   end
-  return color
 end
 
-function util.randomColor(color)
-  if color ~= "NONE" then
-    local hsl = hsluv.hex_to_hsluv(color)
-    hsl[1] = math.random(1, 360)
-    return hsluv.hsluv_to_hex(hsl)
+---@param hi_name string
+---@param hi gt.Highlight|gt.LinkHighlight
+util.highlight = function(hi_name, hi)
+  if hi.fg then
+    util.used_color[hi.fg] = true
   end
-  return color
-end
+  if hi.bg then
+    util.used_color[hi.bg] = true
+  end
+  if hi.sp then
+    util.used_color[hi.sp] = true
+  end
 
-function util.getColor(color)
-  if vim.o.background == "dark" then return color end
-  if vim.o.background == "light" then return color end
-  if not util.colorCache[color] then util.colorCache[color] = util.invertColor(color) end
-  return util.colorCache[color]
-end
+  local style = hi.style and 'gui=' .. hi.style or 'gui=NONE'
+  local fg = hi.fg and 'guifg=' .. hi.fg or 'guifg=NONE'
+  local bg = hi.bg and 'guibg=' .. hi.bg or 'guibg=NONE'
+  local sp = hi.sp and 'guisp=' .. hi.sp or ''
 
--- local ns = vim.api.nvim_create_namespace("github-theme")
-function util.highlight(group, color)
-  if color.fg then util.colorsUsed[color.fg] = true end
-  if color.bg then util.colorsUsed[color.bg] = true end
-  if color.sp then util.colorsUsed[color.sp] = true end
+  local hi_cmd = string.format('highlight %s %s %s %s %s', hi_name, style, fg, bg, sp)
 
-  local style = color.style and "gui=" .. color.style or "gui=NONE"
-  local fg = color.fg and "guifg=" .. util.getColor(color.fg) or "guifg=NONE"
-  local bg = color.bg and "guibg=" .. util.getColor(color.bg) or "guibg=NONE"
-  local sp = color.sp and "guisp=" .. util.getColor(color.sp) or ""
-
-  local hl = "highlight " .. group .. " " .. style .. " " .. fg .. " " .. bg .. " " .. sp
-
-  if color.link then
-    vim.cmd("highlight! link " .. group .. " " .. color.link)
+  if hi.link then
+    vim.cmd(string.format('highlight! link %s %s', hi_name, hi.link))
   else
     -- local data = {}
     -- if color.fg then data.foreground = color.fg end
@@ -99,195 +103,147 @@ function util.highlight(group, color)
     -- if color.sp then data.special = color.sp end
     -- if color.style then data[color.style] = true end
     -- vim.api.nvim_set_hl(ns, group, data)
-    vim.cmd(hl)
+    vim.cmd(hi_cmd)
   end
 end
 
-function util.debug(colors)
-  colors = colors or require("github-theme.colors")
-  -- Dump unused colors
-  for name, color in pairs(colors) do
-    if type(color) == "table" then
-      util.debug(color)
-    else
-      if util.colorsUsed[color] == nil then
-        print("not used: " .. name .. " : " .. color)
-      end
-    end
-  end
-end
-
---- Delete the autocmds when the theme changes to something else
-function util.onColorScheme()
+---Delete the autocmds when the theme changes to something else
+util.on_colorscheme = function()
   if vim.g.colors_name ~= util.colors_name then
-    vim.cmd("autocmd! " .. util.colors_name)
-    vim.cmd("augroup!" .. util.colors_name)
+    vim.cmd('autocmd! ' .. util.colors_name)
+    vim.cmd('augroup!' .. util.colors_name)
   end
 end
 
----@param config github-theme.Config
-function util.autocmds(config)
-  vim.cmd("augroup " .. util.colors_name)
-  vim.cmd([[  autocmd!]])
-  vim.cmd([[  autocmd ColorScheme * lua require("github-theme.util").onColorScheme()]])
+---@param config gt.ConfigSchema
+util.autocmds = function(config)
+  vim.cmd(string.format('augroup %s ', util.colors_name))
+  vim.cmd('autocmd!')
+  vim.cmd('autocmd ColorScheme * lua require("github-theme.util").on_colorscheme()')
   if config.dev then
-    vim.cmd("  autocmd BufWritePost */lua/github-theme/** nested colorscheme " ..
-                util.colors_name)
+    vim.cmd(string.format('autocmd BufWritePost */lua/github-theme/** nested colorscheme %s', util.colors_name))
   end
   for _, sidebar in ipairs(config.sidebars) do
-    if sidebar == "terminal" then
-      vim.cmd(
-          [[  autocmd TermOpen * setlocal winhighlight=Normal:NormalSB,SignColumn:SignColumnSB]])
+    if sidebar == 'terminal' then
+      vim.cmd('autocmd TermOpen * setlocal winhighlight=Normal:NormalSB,SignColumn:SignColumnSB')
     else
-      vim.cmd([[  autocmd FileType ]] .. sidebar ..
-                  [[ setlocal winhighlight=Normal:NormalSB,SignColumn:SignColumnSB]])
+      vim.cmd(string.format('autocmd FileType %s setlocal winhighlight=Normal:NormalSB,SignColumn:SignColumnSB', sidebar))
     end
   end
-  vim.cmd([[augroup end]])
+  vim.cmd('augroup end')
 end
 
--- Simple string interpolation.
---
--- Example template: "${name} is ${value}"
---
----@param str string template string
----@param table table key value pairs to replace in the string
-function util.template(str, table)
-  return (str:gsub("($%b{})", function(w)
-    return table[w:sub(3, -2)] or w
-  end))
+---@param base gt.Highlights.Base
+util.syntax = function(base)
+  for hi_name, hi in pairs(base) do
+    util.highlight(hi_name, hi)
+  end
 end
 
-function util.syntax(syntax)
-  for group, colors in pairs(syntax) do util.highlight(group, colors) end
-end
-
----@param colors github-theme.ColorScheme
-function util.terminal(colors)
+---@param colors gt.ColorPalette
+util.terminal = function(colors)
   -- dark
   vim.g.terminal_color_0 = colors.black
   vim.g.terminal_color_8 = colors.fg_dark
-
   -- light
   vim.g.terminal_color_7 = colors.fg_dark
-  vim.g.terminal_color_15 = colors.term_fg
-
+  vim.g.terminal_color_15 = colors.fg_term
   -- colors
   vim.g.terminal_color_1 = colors.red
   vim.g.terminal_color_9 = colors.bright_red
-
   vim.g.terminal_color_2 = colors.green
   vim.g.terminal_color_10 = colors.bright_green
-
   vim.g.terminal_color_3 = colors.yellow
   vim.g.terminal_color_11 = colors.bright_yellow
-
   vim.g.terminal_color_4 = colors.blue
   vim.g.terminal_color_12 = colors.bright_blue
-
   vim.g.terminal_color_5 = colors.magenta
   vim.g.terminal_color_13 = colors.bright_magenta
-
   vim.g.terminal_color_6 = colors.cyan
   vim.g.terminal_color_14 = colors.bright_cyan
+end
 
-  if vim.o.background == "light" then
-    for i = 0, 15, 1 do
-      vim.g["terminal_color_" .. i] = util.getColor(vim.g["terminal_color_" .. i])
+---@param hi gt.Highlights
+util.load = function(hi)
+  local theme_style = hi.config.theme_style
+  util.colors_name = 'github_' .. theme_style
+
+  --Setting
+  vim.cmd('hi clear')
+  if vim.fn.exists('syntax_on') then
+    vim.cmd('syntax reset')
+  end
+
+  vim.o.termguicolors = true
+  if theme_style == 'light' or theme_style == 'light_default' then
+    vim.o.background = 'light'
+  else
+    vim.o.background = 'dark'
+  end
+  vim.g.colors_name = util.colors_name
+
+  --Load ColorScheme
+  util.syntax(hi.base)
+  util.autocmds(hi.config)
+  util.terminal(hi.colors)
+  util.syntax(hi.plugins)
+end
+
+---Override custom highlights in `group`
+---@param group table
+---@param overrides table
+util.apply_overrides = function(group, overrides)
+  for k, v in pairs(overrides) do
+    if group[k] ~= nil and type(v) == 'table' then
+      group[k] = v
     end
   end
 end
 
----@param theme github-theme.Theme
-function util.load(theme)
-  vim.cmd("hi clear")
-  if vim.fn.exists("syntax_on") then vim.cmd("syntax reset") end
+---@param colors gt.ColorPalette
+---@param oride_colors gt.ColorPalette
+---@return gt.ColorPalette
+function util.color_overrides(colors, oride_colors)
+  if type(oride_colors) == 'table' then
+    for key, value in pairs(oride_colors) do
+      local err_msg = string.format('color "%s" does not exist', key)
 
-  util.colors_name = "github_" .. vim.g.github_theme_style
-
-  vim.o.termguicolors = true
-  vim.g.colors_name = util.colors_name
-  -- vim.api.nvim__set_hl_ns(ns)
-
-  -- load base theme
-  util.syntax(theme.base)
-  util.autocmds(theme.config)
-  util.terminal(theme.colors)
-  util.syntax(theme.plugins)
-
-  vim.defer_fn(function()
-    util.syntax(theme.defer)
-  end, 100)
-end
-
----@param colors github-theme.ColorScheme
----@param config github-theme.Config
-function util.color_overrides(colors, config)
-  if type(config.colors) == "table" then
-    for key, value in pairs(config.colors) do
-      if not colors[key] then error("Color " .. key .. " does not exist") end
-
+      if not colors[key] then
+        error(err_msg)
+      end
       -- Patch: https://github.com/ful1e5/onedark.nvim/issues/6
-      if type(colors[key]) == "table" then
-        util.color_overrides(colors[key], {colors = value})
+      if type(colors[key]) == 'table' then
+        util.color_overrides(colors[key], value)
       else
-        if value:lower() == "none" then
+        if value:lower() == types.gt.HighlightStyle.None:lower() then
           -- set to none
-          colors[key] = "NONE"
-        elseif string.sub(value, 1, 1) == "#" then
+          colors[key] = types.gt.HighlightStyle.None
+        elseif string.sub(value, 1, 1) == '#' then
           -- hex override
           colors[key] = value
         else
           -- another group
-          if not colors[value] then error("Color " .. value .. " does not exist") end
+          if not colors[value] then
+            error(err_msg)
+          end
           colors[key] = colors[value]
         end
       end
     end
   end
+  return colors
 end
 
-function util.light(brightness)
-  for hl_name, hl in pairs(vim.api.nvim__get_hl_defs(0)) do
-    local def = {}
-    for key, def_key in pairs({foreground = "fg", background = "bg", special = "sp"}) do
-      if type(hl[key]) == "number" then
-        local hex = string.format("#%06x", hl[key])
-        local color = util.invertColor(hex)
-        if brightness then color = util.brighten(hex, brightness) end
-        table.insert(def, "gui" .. def_key .. "=" .. color)
-      end
-    end
-    if hl_name ~= "" and #def > 0 then
-      for _, style in pairs({"bold", "italic", "underline", "undercurl", "reverse"}) do
-        if hl[style] then table.insert(def, "gui=" .. style) end
-      end
-
-      vim.cmd("highlight! " .. hl_name .. " " .. table.concat(def, " "))
-    end
-  end
-end
-
-function util.random()
-  local colors = {}
-  for hl_name, hl in pairs(vim.api.nvim__get_hl_defs(0)) do
-    local def = {}
-    for key, def_key in pairs({foreground = "fg", background = "bg", special = "sp"}) do
-      if type(hl[key]) == "number" then
-        local hex = string.format("#%06x", hl[key])
-        local color = colors[hex] and colors[hex] or util.randomColor(hex)
-        colors[hex] = color
-        table.insert(def, "gui" .. def_key .. "=" .. color)
-      end
-    end
-    if hl_name ~= "" and #def > 0 then
-      for _, style in pairs({"bold", "italic", "underline", "undercurl", "reverse"}) do
-        if hl[style] then table.insert(def, "gui=" .. style) end
-      end
-
-      vim.cmd("highlight! " .. hl_name .. " " .. table.concat(def, " "))
-    end
-  end
+---Simple string interpolation.
+---
+---Example template: "${name} is ${value}"
+---
+---@param str string template string
+---@param table table key value pairs to replace in the string
+util.template = function(str, table)
+  return (str:gsub('($%b{})', function(w)
+    return table[w:sub(3, -2)] or w
+  end))
 end
 
 return util